TST: reorganize misplaced month / business month tests (#27085) (#42909)
